health_tracker/requirements.txt
transformers>=4.40.0
torch>=2.0.0
sentencepiece
accelerate
colorama
pytesseract
pillow
pdf2image
python-multipart